mysite/requirements.txt
django==2.2.0
psycopg2==2.8
django-taggit==0.22.2
Markdown==2.6.11
